refactor(session-replay): add injectable mask renderer
