+++ Sources/Swift/Tools/ViewCapture/SentryDefaultViewRenderer.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+@objcMembers
+class SentryDefaultViewRenderer: NSObject, SentryViewRenderer {
+    func render(view: UIView) -> UIImage {
+        let image = UIGraphicsImageRenderer(size: view.bounds.size).image { _ in
+            view.drawHierarchy(in: view.bounds, afterScreenUpdates: false)
+        }
+        return image
+    }
+}
