+++ Sources/Swift/Tools/ViewCapture/SentryMaskRenderer.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+protocol SentryMaskRenderer {
+    func maskScreenshot(screenshot image: UIImage, size: CGSize, masking: [RedactRegion]) -> UIImage
+}
