+++ Sources/Swift/Tools/ViewCapture/SentryViewRenderer.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+@objc protocol SentryViewRenderer {
+    func render(view: UIView) -> UIImage
+}
